intergrate with NGan Luong

+++ TeduShop.Web/Assets/client/js/controllers/shoppingCart.js
@@ -10,11 +10,11 @@
                 address: "required",
                 email: {
                     required: true,
-                    email:true
+                    email: true
                 },
                 phone: {
                     required: true,
-                    number:true
+                    number: true
                 }
             },
             messages: {
@@ -83,9 +83,22 @@
             if (isValid) {
                 cart.createOrder();
             }
-           
+
         });
 
+        $('input[name="paymentMethod"]').off('click').on('click', function () {
+            if ($(this).val() == 'NL') {
+                $('.boxContent').hide();
+                $('#nganluongContent').show();
+            }
+            else if ($(this).val() == 'ATM_ONLINE') {
+                $('.boxContent').hide();
+                $('#bankContent').show();
+            }
+            else {
+                $('.boxContent').hide();
+            }
+        });
     },
     getLoginUser: function () {
         $.ajax({
@@ -111,7 +124,8 @@
             CustomerEmail: $('#txtEmail').val(),
             CustomerMobile: $('#txtPhone').val(),
             CustomerMessage: $('#txtMessage').val(),
-            PaymentMethod: "Thanh toán tiền mặt",
+            PaymentMethod: $('input[name="paymentMethod"]:checked').val(),
+            BankCode: $('input[groupname="bankcode"]:checked').prop('id'),
             Status: false
         }
         $.ajax({
@@ -123,13 +137,23 @@
             },
             success: function (response) {
                 if (response.status) {
-                    console.log('create order ok');
-                    $('#divCheckout').hide();
-                    cart.deleteAll();
-                    setTimeout(function () {
-                        $('#cartContent').html('Cảm ơn bạn đã đặt hàng thành công. Chúng tôi sẽ liên hệ sớm nhất.');
-                    }, 2000);
+                    if (response.urlCheckout != undefined && response.urlCheckout != '') {
+                        
+                        window.location.href = response.urlCheckout;
+                    }
+                    else {
+                        console.log('create order ok');
+                        $('#divCheckout').hide();
+                        cart.deleteAll();
+                        setTimeout(function () {
+                            $('#cartContent').html('Cảm ơn bạn đã đặt hàng thành công. Chúng tôi sẽ liên hệ sớm nhất.');
+                        }, 2000);
+                    }
 
+                }
+                else {
+                    $('#divMessage').show();
+                    $('#divMessage').text(response.message);
                 }
             }
         });
